LayoutTests/http/tests/navigation/sec-fetch-site-header-on-crossorigin-redirection-expected.txt

Tests the behavior of sec-fetch-site header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS secFetchSite is expected
PASS successfullyParsed is true

TEST COMPLETE
